pyproject.toml
[project]
name = "ecommcart-qa"
version = "0.1.0"
description = ""
authors = [
    {name = "arya@189",email = "arya.tamhane@atlascopco.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "behave (>=1.2.6,<2.0.0)",
    "allure-behave (>=2.14.3,<3.0.0)",
    "playwright (>=1.52.0,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
